requirements.txt
absl-py==0.4.1
annoy==1.13.0
anytree==2.6.0
astor==0.7.1
atari-py==0.1.7
certifi==2018.11.29
chardet==3.0.4
Click==7.0
cloudpickle==0.6.1
cycler==0.10.0
dill==0.2.8.2
future==0.17.1
gast==0.2.0
glob2==0.6
grpcio==1.14.2
gym==0.10.5
h5py==2.8.0
idna==2.7
joblib==0.13.0
Keras==2.2.2
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
Markdown==2.6.11
matplotlib==2.2.3
mpi4py==3.0.0
numpy==1.14.5
opencv-python==3.4.4.19
pandas==0.23.4
Pillow==5.2.0
progressbar2==3.38.0
protobuf==3.6.1
pyemd==0.5.1
pyglet==1.3.2
PyOpenGL==3.1.0
pyparsing==2.2.0
-e git+git@github.com:zsoltzombori/leancop_ml.git@84ef5248a3fce0428218829baf06c4ebe24a0422#egg=pyswip&subdirectory=pyswip
python-dateutil==2.7.3
python-utils==2.3.0
pytz==2018.5
PyYAML==3.13
pyzmq==17.1.2
requests==2.20.1
scikit-learn==0.19.2
scipy==1.1.0
seaborn==0.9.0
six==1.11.0
sklearn==0.0
stable-baselines==2.1.1
tensorboard==1.12.0
tensorflow-gpu==1.12.0
termcolor==1.1.0
tqdm==4.28.1
urllib3==1.24.1
Werkzeug==0.14.1
xgboost==0.81
zmq==0.0.0
